feat: #798 - now a unique page for preferences, in accordion (#820)
Deleted file:
* `settings_page.dart`: all the code was moved to `user_preferences_page.dart`

New files:
* `abstract_user_preferences.dart`: Abstraction of a collapsed/expanded display for the preferences page.
* `user_preferences_attribute_group.dart`: Collapsed/expanded display of one attribute group for the preferences page.
* `user_preferences_food.dart`: Collapsed/expanded display of attribute groups for the preferences page.
* `user_preferences_profile.dart`: Collapsed/expanded display of profile for the preferences page.
* `user_preferences_settings.dart`: Collapsed/expanded display of settings for the preferences page.

Impacted files:
* `app_en.arb`: added 6 labels for preferences titles and subtitles
* `app_fr.arb`: added 6 labels for preferences titles and subtitles
* `labeler.yml`: removed reference to `settings_page.dart`
* `user_preferences.dart`: added bool flags getter and setter
* `user_preferences_page.dart`: now displays in accordion the same preferences, plus the former settings_page, plus the former action bar
* `user_preferences_page-blue-dark.png`
* `user_preferences_page-blue-light.png`
* `user_preferences_page-brown-dark.png`
* `user_preferences_page-brown-light.png`
* `user_preferences_page-green-dark.png`
* `user_preferences_page-green-light.png`
