+++ packages/smooth_app/lib/pages/user_preferences_food.dart
@@ -1,0 +1,83 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_gen/gen_l10n/app_localizations.dart';
+import 'package:openfoodfacts/model/AttributeGroup.dart';
+import 'package:smooth_app/data_models/product_preferences.dart';
+import 'package:smooth_app/data_models/user_preferences.dart';
+import 'package:smooth_app/pages/abstract_user_preferences.dart';
+import 'package:smooth_app/pages/user_preferences_attribute_group.dart';
+
+/// Collapsed/expanded display of attribute groups for the preferences page.
+class UserPreferencesFood extends AbstractUserPreferences {
+  UserPreferencesFood({
+    required this.productPreferences,
+    required final Function(Function()) setState,
+    required final BuildContext context,
+    required final UserPreferences userPreferences,
+    required final AppLocalizations appLocalizations,
+    required final ThemeData themeData,
+  }) : super(
+          setState: setState,
+          context: context,
+          userPreferences: userPreferences,
+          appLocalizations: appLocalizations,
+          themeData: themeData,
+        );
+
+  final ProductPreferences productPreferences;
+
+  static const List<String> _ORDERED_ATTRIBUTE_GROUP_IDS = <String>[
+    AttributeGroup.ATTRIBUTE_GROUP_NUTRITIONAL_QUALITY,
+    AttributeGroup.ATTRIBUTE_GROUP_INGREDIENT_ANALYSIS,
+    AttributeGroup.ATTRIBUTE_GROUP_ENVIRONMENT,
+    AttributeGroup.ATTRIBUTE_GROUP_PROCESSING,
+    AttributeGroup.ATTRIBUTE_GROUP_LABELS,
+    AttributeGroup.ATTRIBUTE_GROUP_ALLERGENS,
+  ];
+
+  @override
+  bool isCollapsedByDefault() => false;
+
+  @override
+  String getPreferenceFlagKey() => 'attributes';
+
+  @override
+  Widget getTitle() => Text(
+        appLocalizations.myPreferences_food_title,
+        style: themeData.textTheme.headline2,
+      );
+
+  @override
+  Widget? getSubtitle() => Text(appLocalizations.myPreferences_food_subtitle);
+
+  @override
+  List<Widget> getBody() {
+    final List<AttributeGroup> groups =
+        _reorderGroups(productPreferences.attributeGroups!);
+    final List<Widget> result = <Widget>[];
+    for (final AttributeGroup group in groups) {
+      final AbstractUserPreferences abstractUserPreferences =
+          UserPreferencesAttributeGroup(
+        productPreferences: productPreferences,
+        group: group,
+        setState: setState,
+        context: context,
+        userPreferences: userPreferences,
+        appLocalizations: appLocalizations,
+        themeData: themeData,
+      );
+
+      result.addAll(abstractUserPreferences.getContent());
+    }
+    return result;
+  }
+
+  List<AttributeGroup> _reorderGroups(List<AttributeGroup> groups) {
+    final List<AttributeGroup> result = <AttributeGroup>[];
+    for (final String id in _ORDERED_ATTRIBUTE_GROUP_IDS) {
+      result.addAll(groups.where((AttributeGroup g) => g.id == id));
+    }
+    result.addAll(groups.where(
+        (AttributeGroup g) => _ORDERED_ATTRIBUTE_GROUP_IDS.contains(g.id)));
+    return result;
+  }
+}
